Simplify by combining AgGridWithDetails into one file
